physical-ai-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
qdrant-client==1.7.0
openai==1.3.5
langchain==0.0.341
langchain-community==0.0.28
python-multipart==0.0.6
tiktoken==0.5.1
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2